backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
pydantic
praw
nltk
vaderSentiment
pandas
email-validator
python-multipart
transformers
torch
